fix(odata-service): Removes ui5-config -> odata-service-writer dep
The net result is that we no longer will attempt to use TS 4.1 types
(from yaml) when importing odata-service-writer
