+++ packages/ui5-config/src/index.ts
@@ -1,3 +1,2 @@
 export { UI5Config } from './ui5-config';
 export * from './middlewares';
-export { OdataService, OdataVersion } from './types';
