+++ packages/ui5-config/src/types.ts
@@ -1,24 +1,3 @@
-export enum OdataVersion {
-    v2 = '2',
-    v4 = '4'
-}
-export interface OdataService {
-    url?: string;
-    client?: string;
-    destination?: {
-        name: string;
-        instance?: string;
-    };
-    path?: string;
-    version: OdataVersion;
-    name?: string;
-    model?: string;
-    metadata?: string;
-    annotations?: {
-        technicalName: string;
-        xml: string;
-    };
-}
 
 export interface Destination {
     destination: string;
